修复一些BUG，post的方法待实现
